Reuse the password file if it exists and is valid
If the DB password file saved in root is valid (contains the variable DB_PASSWORD and DB_ROOT_PASSWORD) reuse it rather than generating new passwords. This will make it unnecessary to save new passwords every time a jail is pulled down and rebuilt.
If the password file doesn't exist, newly generated passwords are created and saved in the password file.
If the password file exists, but new passwords are preferred, just delete the password file before running the script.